JSON data files
added JSON data files for posts and accounts

Please add more samples to both posts and accounts up to 5 posts with 1-2 comments each and up to 5 total accounts. Try to make the data as consistent as possible for easy implementation for MCO Phase 2

If you think we need another field for data we're gonna need later, feel free to add. I can't add comments on JSON files so if you have a question regarding the data and format, feel free to ask me